Add Heat template and a playbook that uses it
The Heat template configures an initial version of the stack used for
Etherpad. No configuration is added yet, so the virtual machines in the
stack will be completely unconfigured for now.

Add an example Heat environment file for use with the Heat template and
documentation that has a brief description of what this piece of Ansible
code will do and requirements for running it.