Add Etherpad configuration
Configure Etherpad on the frontend node and connect it to the database
cluster that has been created in a previous stage. Install dependencies
for Etherpad.

Modify the Heat template so that port 9001 is exposed by the frontend
security group.

Also tweak SSH settings a bit by adding ControlMaster config to speed
things up and by adding the StrictHostKeyChecking=no option to
ansible.cfg.