Use 10m instead of 10h for SSH ControlPersist
To avoid issues with ephemeral machines going away when building and
rebuilding the stack, use a 10 minute ControlPersist instead of 10
hours. With 10 hours, a control master socket to a dead machine can stay
on the system for 10 hours, which means if you set up a machine with
that same IP it can't be accessed because SSH will try to use the
existing non-functional master socket.