More detailed documentation and streamlined usage
Document requirements for quota, the OpenStack environment and the
operating systems supported for VMs.

Amend the documentation so that the my-heat-params.yml file is placed
under playbooks/ where the build-heat-stack playbook will be able to
find it directly. Because of this change it is no longer necessary to
specify the location of the Heat parameter file when running the Ansible
code. Modify build-heat-stack.yml so that it has a default location for
the parameter file.

Make it possible to restrict where the Etherpad application will be made
available instead of hardcoding 0.0.0.0/0 in the Heat template. This
introduces an additional mandatory parameter for the user to set. The
parameter is mandatory so that the user needs to consider where to
expose the application and it is not accidentally exposed too widely.