Wait for SSH connectivity
Before moving on to configuration after deploying virtual machines, wait
for the virtual machines to become available. This worked previously
because VM provisioning was so fast, but if there is any delay then this
can lead to errors if Ansible doesn't make sure SSH is available before
trying to manage the VMs.